refactor: use EntityTable for invoices Removes InvoiceTableYear Removes InvoiceTable Provides single EntityTable in InvoiceOverview
